+++ src/resources/views/emails/tailwind.blade.php
@@ -1,1 +1,0 @@
-@include('laravelexceptionnotifier::emails.modern', ['framework' => 'tailwind'])